+++ pom.xml
@@ -4,9 +4,9 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>de.ufomc</groupId>
+    <groupId>dev.ufo</groupId>
     <artifactId>ufodata</artifactId>
-    <version>0.2.0-SNAPSHOT</version>
+    <version>0.3.1</version>
 
     <properties>
         <maven.compiler.source>21</maven.compiler.source>
