--- src/main/java/de/ufomc/config/pre/TypeValue.java
@@ -1,15 +1,0 @@
-package de.ufomc.config.pre;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class TypeValue {
-
-    private String type;
-    private Object value;
-
-}
