+++ src/main/java/dev/ufo/ufodata/io/UfoFile.java
@@ -1,12 +1,12 @@
-package de.ufomc.config.io;
+package dev.ufo.ufodata.io;
 
-import de.ufomc.config.checks.CheckType;
-import de.ufomc.config.core.QueuedAsyncExecution;
-import de.ufomc.config.format.JsonFormatter;
-import de.ufomc.config.format.ListFormatter;
-import de.ufomc.config.format.MapFormatter;
-import de.ufomc.config.format.ObjectFormatter;
-import de.ufomc.config.pre.TypeValue;
+import dev.ufo.ufodata.checks.CheckType;
+import dev.ufo.ufodata.core.QueuedAsyncExecution;
+import dev.ufo.ufodata.format.JsonFormatter;
+import dev.ufo.ufodata.format.ListFormatter;
+import dev.ufo.ufodata.format.MapFormatter;
+import dev.ufo.ufodata.format.ObjectFormatter;
+import dev.ufo.ufodata.pre.TypeValue;
 import lombok.Getter;
 
 import java.io.File;
